Fix balance panel
