+++ src/components/Tables/Account/AccountOutgoingStreamsTable.tsx
@@ -38,7 +38,7 @@ import { sfSubgraph } from "../../../redux/store";
 import AccountAddress from "../../AccountAddress";
 import ClearInputAdornment from "../../ClearInputAdornment";
 import DetailsButton from "../../DetailsButton";
-import TokenBalance from "../../FlowingBalanceWithToken";
+import FlowingBalanceWithToken from "../../FlowingBalanceWithToken";
 import FlowRate from "../../FlowRate";
 import InfinitePagination from "../../InfinitePagination";
 import InfoTooltipBtn from "../../InfoTooltipBtn";
@@ -377,7 +377,7 @@ const AccountOutgoingStreamsTable: FC<AccountOutgoingStreamsTableProps> = ({
                 <FlowRate flowRate={stream.currentFlowRate} />
               </TableCell>
               <TableCell data-cy={"outgoing-total-streamed"}>
-                <TokenBalance
+                <FlowingBalanceWithToken
                   balance={stream.streamedUntilUpdatedAt}
                   balanceTimestamp={stream.updatedAtTimestamp}
                   flowRate={stream.currentFlowRate}
