+++ src/pages/[_network]/streams/[_id].tsx
@@ -26,7 +26,7 @@ import AccountAddress from "../../../components/AccountAddress";
 import AppLink from "../../../components/AppLink";
 import CopyLink from "../../../components/CopyLink";
 import FlowingBalance from "../../../components/FlowingBalance";
-import TokenBalance from "../../../components/FlowingBalanceWithToken";
+import FlowingBalanceWithToken from "../../../components/FlowingBalanceWithToken";
 import FlowRate from "../../../components/FlowRate";
 import InfoTooltipBtn from "../../../components/InfoTooltipBtn";
 import SkeletonAddress from "../../../components/skeletons/SkeletonAddress";
@@ -249,7 +249,7 @@ export const StreamPageContent: FC<{ streamId: string; network: Network }> = ({
               primary={
                 stream ? (
                   <>
-                    <TokenBalance
+                    <FlowingBalanceWithToken
                       balance={stream.streamedUntilUpdatedAt}
                       balanceTimestamp={stream.updatedAtTimestamp}
                       flowRate={stream.currentFlowRate}
